fse-828: 상미기한별 재고 스케쥴러를 통한 저장로직 완성

+++ src/main/kotlin/com/smartfoodnet/fninventory/stock/entity/StockByBestBefore.kt
@@ -32,7 +32,7 @@ class StockByBestBefore(
     var shippingProductId: Long,
 
     @Column(name = "best_before")
-    var bestBefore: Int? = null,
+    var bestBefore: Float? = null,
 
     @Column(name = "manufacture_date")
     var manufactureDate: LocalDateTime? = null,
